shells/bash: symlink /bin/bash -> /usr/local/bin/bash
Using "#!/bin/bash" as shebang in shell scripts is rather common these
days. And while it may be argued that programmers should not hard-code
the path to bash in the shebang, it doesn't change the fact that Linux
is so obiquitous these days that most developers will not think twice
about doing so.

This means that it becomes challenging to run bash scripts under
FreeBSD. Especially if a set of bash scripts are calling each other,
there isn't really a workaround.

Signed-off-by: Markus Mayer <code@mmayer.net>